Move NameShrinker out of registry and add container methods.
